py/discord-bot/wro-bot/requirements.txt
git+https://github.com/zawodev/msgcount.py@rewrite
discord.py
pytz
python-dotenv
requests
validators
asyncio
uuid
datetime